# Make cloning git repos faster