## What was the end-user or developer problem that led to this PR?

The problem was that cloning bit git repos can be very slow, because we are cloning the full history of the repo.

## What is your fix for the problem, implemented in this PR?

My fix is to use `--depth 1` for remote clones to only download what's necessary.

This approach doesn't really work and needs more work. A potential fix was suggested [here](https://github.com/rubygems/rubygems/issues/3332#issuecomment-652990815).

Fixes #3332.
Fixes #3378.
Fixes #3775.
Supersedes https://github.com/rubygems/bundler/pull/7054.

## Make sure the following tasks are checked

- [x] Describe the problem / feature
- [ ] Write [tests](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#tests) for features and bug fixes
- [ ] Write code to solve the problem
- [ ] Make sure you follow the [current code style](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#code-formatting) and [write meaningful commit messages without tags](https://github.com/rubygems/rubygems/blob/master/bundler/doc/development/PULL_REQUESTS.md#commit-messages)