@deivid-rodriguez 2021-03-24
This should be mostly working now, and it does result in a significant improvement for large git repositories. For example, bundling rails edge with a cold cache:

#### Before

```
real 0m26,123s
user 0m39,029s
sys 0m4,780s

248M /home/deivid/.rbenv/versions/3.0.0/lib/ruby/gems/3.0.0/cache/bundler/git/rails-fcf0202857b07db1a0f6220dae5ca99319ca0f32
290M /home/deivid/.rbenv/versions/3.0.0/lib/ruby/gems/3.0.0/bundler/gems/rails-e287eed77a2a
```

#### After

```
real 0m6,326s
user 0m3,572s
sys 0m0,639s

9,1M /home/deivid/.rbenv/versions/3.0.0/lib/ruby/gems/3.0.0/cache/bundler/git/rails-fcf0202857b07db1a0f6220dae5ca99319ca0f32
51M /home/deivid/.rbenv/versions/3.0.0/lib/ruby/gems/3.0.0/bundler/gems/rails-e287eed77a2a
```

I did try to make this backwards compatible but I'm afraid this needs more work because this feature of git depends on both client and server versions, and also on the server having the `allow-reachable-sha1-in-want` capability enabled.

So we probably need to ask this to the server to decide whether we should fetch the whole repository or not. Apparently git servers have this `/info/refs?service=git-upload-pack` that provides this information:

```
 $ curl https://github.com/git/git/info/refs\?service=git-upload-pack -sSL -o - | grep -a allow-reachable-sha1-in-want
00000156142430338477d9d1bb25be66267225fb58498d92 HEADmulti_ack thin-pack side-band side-band-64k ofs-delta shallow deepen-since deepen-not deepen-relative no-progress include-tag multi_ack_detailed allow-tip-sha1-in-want allow-reachable-sha1-in-want no-done symref=HEAD:refs/heads/master filter object-format=sha1 agent=git/github-g9c23973022a1
 ```