@deivid-rodriguez 2021-09-04
Hi. No updates. What's left to do is what I said in the previous message. Alternatively, instead of try to explicitly "feature detect" this capability, we could rescue errors when fetching specific revisions, and fallback to the current mode in that case.