@mitchellhenke 2021-09-16
I think this would be super!

I don't want to expand the scope or delay this work, but wanted to mention that a couple related options to minimize clone size are [--branch](https://git-scm.com/docs/git-clone/2.16.6#Documentation/git-clone.txt---branchltnamegt) and [--no-tags](https://git-scm.com/docs/git-clone/2.16.6#Documentation/git-clone.txt---no-tags). My understanding is `--branch` supports branches and tags, but not arbitrary refs.

In terms of feature detection, is there a minimum git version that bundler aims to support? I think all of the clone options mentioned are available on the versions of `git` currently receiving at least security patches.

Thanks for this work, it'd be great to see it land, and I'd be glad to test or help 🙂