@deivid-rodriguez 2021-09-21
The problem is this not only depends on `git` client's version, which is very easy to check, but on git server's version and whether it has the `allow-reachable-sha1-in-want` capability enabled. But perhaps this is an overkill concern and most git servers have this. Maybe we can at least enable this for github.com, gitlab.com or others that we know can properly handle this.