@deivid-rodriguez 2022-12-02
If CI agrees, I believe this is ready.

I decided to pass on supporting git servers without the `allow-reachable-sha1-in-want` capability enabled. If it turns out someone needs that, we can follow up later.

I am still falling back to full clones on some edge cases, and when old client git version is used. The old client code paths are untested, but they are "duplicated" with other tested cases like like pinning to shortened sha's, which are tested, so I believe it's fine.

These are updated results on my laptop for bundling rails edge with a cold cache:

### On master

```
$ rm -rf ~/.asdf/installs/ruby/3.1.3/lib/ruby/gems/3.1.0/cache/bundler/ ~/.asdf/installs/ruby/3.1.3/lib/ruby/gems/3.1.0/bundler Gemfile.lock && time bundle _2.4.0.dev_ lock && du -hs ~/.asdf/installs/ruby/3.1.3/lib/ruby/gems/3.1.0/cache/bundler/ && du -hs /Users/deivid/.asdf/installs/ruby/3.1.3/lib/ruby/gems/3.1.0/bundler
Fetching https://github.com/rails/rails.git
Resolving dependencies...Fetching gem metadata from https://rubygems.org/.......

Writing lockfile to /Users/deivid/Code/playground/rails-edge/Gemfile.lock
bundle _2.4.0.dev_ lock  12,67s user 3,92s system 95% cpu 17,316 total
278M	/Users/deivid/.asdf/installs/ruby/3.1.3/lib/ruby/gems/3.1.0/cache/bundler/
319M	/Users/deivid/.asdf/installs/ruby/3.1.3/lib/ruby/gems/3.1.0/bundler
```

### On this branch

```
$ rm -rf ~/.asdf/installs/ruby/3.1.3/lib/ruby/gems/3.1.0/cache/bundler/ ~/.asdf/installs/ruby/3.1.3/lib/ruby/gems/3.1.0/bundler Gemfile.lock && time bundle _2.4.0.dev_ lock && du -hs ~/.asdf/installs/ruby/3.1.3/lib/ruby/gems/3.1.0/cache/bundler/ && du -hs /Users/deivid/.asdf/installs/ruby/3.1.3/lib/ruby/gems/3.1.0/bundler
Fetching https://github.com/rails/rails.git
Resolving dependencies...Fetching gem metadata from https://rubygems.org/.......

Writing lockfile to /Users/deivid/Code/playground/rails-edge/Gemfile.lock
bundle _2.4.0.dev_ lock  1,25s user 0,69s system 40% cpu 4,804 total
9,5M	/Users/deivid/.asdf/installs/ruby/3.1.3/lib/ruby/gems/3.1.0/cache/bundler/
 49M	/Users/deivid/.asdf/installs/ruby/3.1.3/lib/ruby/gems/3.1.0/bundler
```