@deivid-rodriguez 2022-12-04
I realized that while the previous implementation was indeed making the case of a big single git repo (like pointing to Rails edge) faster, it was actually regressing performance when pointing to many small git repos (like https://github.com/ManageIQ/manageiq).

The reason was that it was using an extra `git fetch` step after `git clone` with `depth=1` in order to be able to fetch the requested ref. This step was not previously necessary when doing a full clone, and it resulted in performance actually worse for small repos.

Now I'm leveraging the `--branch` option to `git clone` mentioned by @mitchellhenke at https://github.com/rubygems/rubygems/pull/4475#issuecomment-920490715, and the current solution now also improves performance for a case like manageiq Gemfile:

### On master

```
$ rm -rf ~/.asdf/installs/ruby/3.0.4/lib/ruby/gems/3.0.0/cache/bundler/ ~/.asdf/installs/ruby/3.0.4/lib/ruby/gems/3.0.0/bundler Gemfile.lock && time bundle _2.4.0.dev_ lock && du -hs ~/.asdf/installs/ruby/3.0.4/lib/ruby/gems/3.0.0/cache/bundler/ && du -hs /Users/deivid/.asdf/installs/ruby/3.0.4/lib/ruby/gems/3.0.0/bundler
Fetching https://github.com/ManageIQ/manageiq-ui-classic
Fetching https://github.com/ManageIQ/manageiq-decorators
Fetching https://github.com/ManageIQ/manageiq-consumption
Fetching https://github.com/ManageIQ/manageiq-content
Fetching https://github.com/ManageIQ/manageiq-api
Fetching https://github.com/ManageIQ/manageiq-automation_engine
Fetching https://github.com/ManageIQ/manageiq-providers-vmware
Fetching https://github.com/ManageIQ/manageiq-providers-scvmm
Fetching https://github.com/ManageIQ/manageiq-providers-ovirt
Fetching https://github.com/ManageIQ/manageiq-providers-openstack
Fetching https://github.com/ManageIQ/manageiq-providers-openshift
Fetching https://github.com/ManageIQ/manageiq-providers-red_hat_virtualization
Fetching https://github.com/ManageIQ/manageiq-providers-redfish
Fetching https://github.com/ManageIQ/manageiq-providers-oracle_cloud
Fetching https://github.com/ManageIQ/manageiq-providers-nuage
Fetching https://github.com/ManageIQ/manageiq-providers-nsxt
Fetching https://github.com/ManageIQ/manageiq-providers-lenovo
Fetching https://github.com/ManageIQ/manageiq-providers-kubevirt
Fetching https://github.com/ManageIQ/manageiq-providers-kubernetes
Fetching https://github.com/ManageIQ/manageiq-providers-ibm_terraform
Fetching https://github.com/ManageIQ/manageiq-providers-ibm_power_vc
Fetching https://github.com/ManageIQ/manageiq-providers-ibm_power_hmc
Fetching https://github.com/ManageIQ/manageiq-providers-ibm_cloud
Fetching https://github.com/ManageIQ/manageiq-providers-ibm_cic
Fetching https://github.com/ManageIQ/manageiq-providers-google
Fetching https://github.com/ManageIQ/manageiq-providers-foreman
Fetching https://github.com/ManageIQ/manageiq-providers-cisco_intersight
Fetching https://github.com/ManageIQ/manageiq-providers-azure_stack
Fetching https://github.com/ManageIQ/manageiq-providers-azure
Fetching https://github.com/ManageIQ/manageiq-providers-awx
Fetching https://github.com/ManageIQ/manageiq-providers-autosde
Fetching https://github.com/ManageIQ/manageiq-providers-ansible_tower
Fetching https://github.com/ManageIQ/amazon_ssa_support.git
Fetching https://github.com/ManageIQ/manageiq-providers-amazon
Fetching https://github.com/ManageIQ/manageiq-schema
Fetching https://github.com/ManageIQ/manageiq-gems-pending.git
Fetching source index from https://rubygems.manageiq.org/
Fetching gem metadata from https://rubygems.org/......
Fetching source index from https://rubygems.org/
Resolving dependencies...
Writing lockfile to /Users/deivid/Code/playground/manageiq/Gemfile.lock
bundle _2.4.0.dev_ lock  32,10s user 10,05s system 50% cpu 1:23,08 total
339M	/Users/deivid/.asdf/installs/ruby/3.0.4/lib/ruby/gems/3.0.0/cache/bundler/
611M	/Users/deivid/.asdf/installs/ruby/3.0.4/lib/ruby/gems/3.0.0/bundler
```

### On this branch

```
$ rm -rf ~/.asdf/installs/ruby/3.0.4/lib/ruby/gems/3.0.0/cache/bundler/ ~/.asdf/installs/ruby/3.0.4/lib/ruby/gems/3.0.0/bundler Gemfile.lock && time bundle _2.4.0.dev_ lock && du -hs ~/.asdf/installs/ruby/3.0.4/lib/ruby/gems/3.0.0/cache/bundler/ && du -hs /Users/deivid/.asdf/installs/ruby/3.0.4/lib/ruby/gems/3.0.0/bundler
Fetching https://github.com/ManageIQ/manageiq-ui-classic
Fetching https://github.com/ManageIQ/manageiq-decorators
Fetching https://github.com/ManageIQ/manageiq-consumption
Fetching https://github.com/ManageIQ/manageiq-content
Fetching https://github.com/ManageIQ/manageiq-api
Fetching https://github.com/ManageIQ/manageiq-automation_engine
Fetching https://github.com/ManageIQ/manageiq-providers-vmware
Fetching https://github.com/ManageIQ/manageiq-providers-scvmm
Fetching https://github.com/ManageIQ/manageiq-providers-ovirt
Fetching https://github.com/ManageIQ/manageiq-providers-openstack
Fetching https://github.com/ManageIQ/manageiq-providers-openshift
Fetching https://github.com/ManageIQ/manageiq-providers-red_hat_virtualization
Fetching https://github.com/ManageIQ/manageiq-providers-redfish
Fetching https://github.com/ManageIQ/manageiq-providers-oracle_cloud
Fetching https://github.com/ManageIQ/manageiq-providers-nuage
Fetching https://github.com/ManageIQ/manageiq-providers-nsxt
Fetching https://github.com/ManageIQ/manageiq-providers-lenovo
Fetching https://github.com/ManageIQ/manageiq-providers-kubevirt
Fetching https://github.com/ManageIQ/manageiq-providers-kubernetes
Fetching https://github.com/ManageIQ/manageiq-providers-ibm_terraform
Fetching https://github.com/ManageIQ/manageiq-providers-ibm_power_vc
Fetching https://github.com/ManageIQ/manageiq-providers-ibm_power_hmc
Fetching https://github.com/ManageIQ/manageiq-providers-ibm_cloud
Fetching https://github.com/ManageIQ/manageiq-providers-ibm_cic
Fetching https://github.com/ManageIQ/manageiq-providers-google
Fetching https://github.com/ManageIQ/manageiq-providers-foreman
Fetching https://github.com/ManageIQ/manageiq-providers-cisco_intersight
Fetching https://github.com/ManageIQ/manageiq-providers-azure_stack
Fetching https://github.com/ManageIQ/manageiq-providers-azure
Fetching https://github.com/ManageIQ/manageiq-providers-awx
Fetching https://github.com/ManageIQ/manageiq-providers-autosde
Fetching https://github.com/ManageIQ/manageiq-providers-ansible_tower
Fetching https://github.com/ManageIQ/amazon_ssa_support.git
Fetching https://github.com/ManageIQ/manageiq-providers-amazon
Fetching https://github.com/ManageIQ/manageiq-schema
Fetching https://github.com/ManageIQ/manageiq-gems-pending.git
Fetching source index from https://rubygems.manageiq.org/
Fetching gem metadata from https://rubygems.org/......
Fetching source index from https://rubygems.org/
Resolving dependencies...
Writing lockfile to /Users/deivid/Code/playground/manageiq/Gemfile.lock
bundle _2.4.0.dev_ lock  16,17s user 6,73s system 31% cpu 1:12,29 total
 30M	/Users/deivid/.asdf/installs/ruby/3.0.4/lib/ruby/gems/3.0.0/cache/bundler/
300M	/Users/deivid/.asdf/installs/ruby/3.0.4/lib/ruby/gems/3.0.0/bundler
```